nested decision

+++ basics/decisions/simple_decision/comparison_operators.py
@@ -1,7 +1,7 @@
 print("enter your first number")
-num1=int(input())
+num1=float(input())
 print("enter your second number")
-num2=int(input())
+num2=float(input())
 if num1>num2:
     print ("first number is bigger")
 elif num2>num1:
